fix: convert PiD output to float32 before returning to ComfyUI
ComfyUI's save_images node calls .cpu().numpy() which does not support
bfloat16. Convert to float32 in pid_image_to_comfy before returning.

+++ core/pid_model_manager.py
@@ -201,7 +201,7 @@ def pid_image_to_comfy(image: torch.Tensor) -> torch.Tensor:
         image = image.squeeze(2)
     if image.dim() != 4:
         raise ValueError(f"Expected image [B,C,H,W], got shape {image.shape}")
-    return ((image + 1.0) / 2.0).clamp(0.0, 1.0).permute(0, 2, 3, 1)
+    return ((image + 1.0) / 2.0).clamp(0.0, 1.0).permute(0, 2, 3, 1).float()
 
 
 def pid_latent_to_comfy(latent: torch.Tensor) -> dict:
